+++ node/resolvers/search/index.ts
@@ -1,5 +1,5 @@
 import { NotFoundError, UserInputError, createMessagesLoader } from '@vtex/api'
-import { pathOr, test } from 'ramda'
+import { pathOr } from 'ramda'
 
 import {
   buildAttributePath,
@@ -535,13 +535,6 @@ export const queries = {
   },
 
   searchMetadata: async (_: any, args: SearchMetadataArgs, ctx: Context) => {
-    const queryTerm = args.query
-    if (queryTerm == null || test(/[?&[\]=]/, queryTerm)) {
-      throw new UserInputError(
-        `The query term contains invalid characters. query=${queryTerm}`
-      )
-    }
-
     if (args.selectedFacets) {
       const { maps, queries } = args.selectedFacets.reduce(
         (acc, { key, value }) => {